MRF/ALE (A. Gardi) -------------------------------------------
* merged MRF_TRANS_CST and MRF_TRANS_LIN in a single MRF_TRANS_LIN
(BLOCK:MRF TYPE=TRANSLATION) since the first case can be easily
handled as a subcase of the 2nd case with no computational burden
(it's pretty fast already)

* created MRF_COMB_OSC that handles both translational AND rotational
oscillations, with possibly different periods, amplitudes and phases
(and of course an independent translational oscillation direction)

* completed mrfvel_abs2rel, mrfvel_rel2abs, mrfpos_rel2abs
subroutines that now also take into account the CENTER_VELOCITY (ignored
before) and work for ANY case in the right way.

* introduced compulsory keys
  . INPUT=RELATIVE/ABSOLUTE 
  . OUTPUT=RELATIVE/(ABSOLUTE) 
The INPUT is meant for BOTH the INITIAL conditions and the BOUNDARY conditions. 

Further development:
  The only incomplete feature is the ABSOLUTE OUTPUT: must choose which is the preferred way
  . create a new separate vertex array (dynamically allocated) with the
    updated absolute positions, and modify all the output routines to
    account for that possible alternate output
  . trick all the output subroutines by making the original vertex
    array to temporary point to the updated absolute positions array and
    then repoint it back to the original array after output

* Changes in keywords of RPMBLOCK:MRF keywords
  OSC_ANGLE   -->  ROT_AMPL
  OSC_PERIOD  -->  ROT_PERIOD
  PHI         -->  ROT_PHI